requirements.txt
Flask>=2.2.0
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
pika==1.3.1
uuid==1.30
Werkzeug==2.3.4
Jinja2==3.1.2
requests>=2.28.1
kafka-python>=2.0.2
pytz>=2022.1
python-dotenv>=0.21.0